+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.30
 
+- Add more status in problem info page
 - Fix solution infinity loading
 - Log websocket messages